Fixed the last rarity in the rarity display as well as the later rarities being unobtainable

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/Item.java
@@ -229,7 +229,7 @@ public class Item implements Bundlable {
 
 	public void randomizeRarity() {
 		rarity = Rarity.NONE;
-		float random = Dungeon.Float(1, Dungeon.LuckDirection.DOWN);
+		double random = Dungeon.Double(1, Dungeon.LuckDirection.DOWN);
 		for (Rarity r : Rarity.values()) {
 			if (random <= r.chance) {
 				rarity = r;
